Refactors custom helpers.
